+++ rustcw/src/bin/rustc-wrapper-difftests-workspace.rs
@@ -1,5 +1,0 @@
-use std::process::ExitCode;
-
-fn main() -> std::io::Result<ExitCode> {
-    rustc_wrapper_difftests::rustc_wrapper_impl()
-}